refactor: 기능별 예외처리 구분
