+++ application/main-application/src/main/java/org/mainapplication/global/error/CustomException.java
@@ -1,15 +1,15 @@
-package org.mainapplication.global.error.exception;
+package org.mainapplication.global.error;
 
 import java.io.Serializable;
 
 import lombok.Getter;
 
 @Getter
-public class CustomException extends RuntimeException implements Serializable {
+public abstract class CustomException extends RuntimeException implements Serializable {
 
 	private final transient ErrorCodeStatus errorCodeStatus;
 
-	public CustomException(ErrorCodeStatus errorCodeStatus) {
+	protected CustomException(ErrorCodeStatus errorCodeStatus) {
 		super(errorCodeStatus.getMessage());
 		this.errorCodeStatus = errorCodeStatus;
 	}
