+++ application/main-application/src/main/java/org/mainapplication/global/error/ErrorCodeStatus.java
@@ -1,4 +1,4 @@
-package org.mainapplication.global.error.exception;
+package org.mainapplication.global.error;
 
 import org.springframework.http.HttpStatus;
