+++ application/main-application/src/main/java/org/mainapplication/global/error/code/CommonErrorCode.java
@@ -1,5 +1,6 @@
-package org.mainapplication.global.error.exception;
+package org.mainapplication.global.error.code;
 
+import org.mainapplication.global.error.ErrorCodeStatus;
 import org.springframework.http.HttpStatus;
 
 import lombok.Getter;
@@ -7,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ErrorCode implements ErrorCodeStatus {
+public enum CommonErrorCode implements ErrorCodeStatus {
 	// 서버 오류
 	METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "지원하지 않는 HTTP method 입니다."),
 	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 오류, 관리자에게 문의하세요"),
