+++ application/main-application/src/main/java/org/mainapplication/global/error/exception/CommonException.java
@@ -1,0 +1,10 @@
+package org.mainapplication.global.error.exception;
+
+import org.mainapplication.global.error.CustomException;
+import org.mainapplication.global.error.ErrorCodeStatus;
+
+public class CommonException extends CustomException {
+	public CommonException(ErrorCodeStatus errorCodeStatus) {
+		super(errorCodeStatus);
+	}
+}
