+++ application/main-application/src/main/java/org/mainapplication/global/oauth2/handler/CustomOAuth2FailureHandler.java
@@ -3,13 +3,9 @@ package org.mainapplication.global.oauth2.handler;
 import java.io.IOException;
 
 import org.mainapplication.global.constants.UrlConstants;
-import org.mainapplication.global.error.ErrorResponse;
-import org.mainapplication.global.response.GlobalResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.stereotype.Component;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
